+++ NEWS.md
@@ -1,4 +1,5 @@
 * ncbi_snp_query(): enable allele frequency for different reference populations, output now a tibble (#97).
+* ncbi_snp_query(): replace calls to RJSONIO with equivalent in jsonlite (#98).
 
 rsnps 0.4.0
 ===========